ngfw-15147 handled save button based on isDirty flag
