+++ untangle-vue-ui/source/src/components/settings/settingsMixin.js
@@ -5,11 +5,8 @@ export default {
   props: {
     settings: { type: [Object, Array], default: null },
     disabled: { type: Boolean, default: false },
-    // matches styles to ETM, uses h1 for page titles and makes the component expand the whole width
     classicView: { type: Boolean, required: false, default: false },
-    // used to show / hide the description field; shown for mfw-ui but hidden in ETM
     showDescription: { type: Boolean, default: false },
-    // keeping isChanged separate from isDirty until we make CD-5042 change for all appliance components
     isChanged: { type: Boolean, default: false },
   },
